\name{Lei301}
\alias{Lei301}

\docType{data}

\title{Ring widths and stem analyses of tree 1 of the Leindefelde 30 stand (Germany)}

\description{
  This data set gives the ring-widths measured on the cross-sections on the main stem and on its large branch for tree 1 of Leienfelde-30 stand.
Ring-widths are given in 1/1000 mm.
\code{Basic.density}, and \code{carbon.content} were measured on wood subsamples. The \code{volume.shrink} coefficient is taken from literature.
}

\usage{data(Lei301)}

\format{A list of data.frames from ring-width tables to stem analysis tables}

\source{Rings were dated and measured in October 2001, last ring is dated 2000, on a LINTAB 3 measuring device linked to the TSAP software program (Frank Rinn, Germany), by Marco Bascietto.
}

\references{Nardi Berti R., La struttura anatomica del legno ed il riconoscimento dei legnami italiani di più comune impiego, Contributi Scientifico-Pratici per una migliore conoscenza ed utilizzazione del legno, CNR 24 (1979) 117-155.
in  Bascietto M., Scarascia-Mugnozza G., A collection of functions to determine annual tree Carbon increment via stem analysis, Ann. Sci. For. (accepted)
}

\author{Marco Bascietto}

\keyword{datasets}
